requirements.txt
paho-mqtt>=1.6.0
pytest>=8.0.0
pytest-mock>=3.14.0
pytest-cov>=6.0.0
pytest-timeout>=2.3.0 